Extract /search/... to separate module
